feat: call add doc on dialog save
