+++ src/App.tsx
@@ -1,16 +1,31 @@
 import { CButton, CContainer } from '@coreui/react';
-import { useState } from 'react';
+import { getDocs } from 'firebase/firestore';
+import { useEffect, useState } from 'react';
 import Header from './components/header';
 import Dialog from './components/record-dialog';
 import Table from './components/table';
+import userCollection from './services/firebase.config';
 
 
 function App() {
   const [dialogOpen, setDialogOpen] = useState<boolean>(false);
+  const [users, setUsers] = useState<Record<string, User>>({});
+
+  useEffect(() => {
+    const getUsers = async () => {
+      const docs = await getDocs(userCollection);
+      const retrievedUsers: Record<string, User> = {};
+      docs.forEach(doc => {
+        retrievedUsers[doc.id] = doc.data();
+      });
+      setUsers(retrievedUsers);
+    }
+    getUsers();
+  }, []);
 
   const handleAddUserClick = () => setDialogOpen(true);
 
-  const renderRecordDialog = (): JSX.Element => <Dialog isOpen={dialogOpen} setOpen={setDialogOpen} isCreate={true} />
+  const renderRecordDialog = (): JSX.Element => <Dialog isOpen={dialogOpen} setOpen={setDialogOpen} users={users} setUsers={setUsers} isCreate={true} />
 
   return (
     <>
@@ -20,7 +35,7 @@ function App() {
           + Add User
         </CButton>
         {renderRecordDialog()}
-        <Table />
+        <Table users={users} />
       </CContainer>
     </>
   );
